requirements.txt
django
django-polymorphic==4.1.0
pillow==11.3.0

